Fix Circular import; Create ltcui.py file
LTCBackend having access to setting the StatusDisplay State was causing
a circular import. In order to prevent future circular imports, and to
clean up ltc.py, I created ltcui.py to hold ui elements.
